% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_individual_clustering.R
\name{apply_BayesSpace}
\alias{apply_BayesSpace}
\title{platform = "Visium" or "ST"}
\usage{
apply_BayesSpace(
  Seurat.data,
  platform = "Visium",
  n.PCs = 15,
  n.HVGs = 2000,
  n.setting = 7,
  nrep = 50000
)
}
\description{
platform = "Visium" or "ST"
}
