% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_individual_clustering.R
\name{apply_DR.SC}
\alias{apply_DR.SC}
\title{platform: 'Visium', "ST", "Other_SRT", "scRNAseq"}
\usage{
apply_DR.SC(
  Seurat.data,
  n_SVGs = 600,
  latent_q = 15,
  num.ct = 7,
  platform = "Visium"
)
}
\description{
platform: 'Visium', "ST", "Other_SRT", "scRNAseq"
}
