% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{clust2Mat}
\alias{clust2Mat}
\title{This function will transfer the spots label to binary similarity matrix}
\usage{
clust2Mat(memb)
}
\arguments{
\item{memb}{the spots label vector}
}
\value{
a binary spots similarity matrix
}
\description{
This function will transfer the spots label to binary similarity matrix
}
