% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{data_process}
\alias{data_process}
\title{This function focuses on cleaning SRT data
and store the SRT expression data, spatial location and image path in the Seurat object}
\usage{
data_process(counts_path, loc_path, img_path, n_HVG = 2000, n_PCA = 20)
}
\arguments{
\item{counts_path}{the .txt path for SRT gene expression data. The file need include gene names and spots names.}

\item{loc_path}{the .txt path for meta data of the spots.}

\item{img_path}{the .tif path of H\& image of the corresponding sequencing tissue.}

\item{n_HVG}{the number of Highly Variation Genes (HVGs) selected by Seurat.}

\item{n_PCA}{the number of low-dimensional of SRT gene expression.}
}
\value{
a Seurat object.
}
\description{
This function focuses on cleaning SRT data
and store the SRT expression data, spatial location and image path in the Seurat object
}
