% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ct_inference.R
\name{data_process_rctd}
\alias{data_process_rctd}
\title{The function is used for the process of scRNA-seq and spatial gene expression data for the input for deconvolution}
\usage{
data_process_rctd(
  sc_exp,
  sc_label,
  spot_exp,
  spot_loc,
  gene_det_in_min_cells_per = 0.01,
  expression_threshold = 1,
  nUMI = 100,
  verbose = FALSE,
  plot = FALSE
)
}
\arguments{
\item{sc_exp}{a matrix where row represents genes and column represents cells}

\item{sc_label}{the cell label of sc_exp}

\item{spot_exp}{a matrix where row represents genes and colum represents the spots.}

\item{spot_loc}{the location of spots}

\item{gene_det_in_min_cells_per}{the min percentage of gene expression in cells.}

\item{expression_threshold}{the min of total expression of genes in all cells.}

\item{nUMI}{the min of total expression of cells in all genes.}

\item{verbose}{whether to output the processing information.}

\item{plot}{whether to plot}
}
\description{
The function is used for the process of scRNA-seq and spatial gene expression data for the input for deconvolution
}
