% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LocalG_spa.R
\name{LocalG_spa}
\alias{LocalG_spa}
\title{Local Getis and Ord's Gi for measure the spatial autocorrelation of selected gene}
\usage{
LocalG_spa(Seurat_data, gene_name, k = 6)
}
\arguments{
\item{Seurat_data}{a SeuratObject created by data_process function in EnSDD R package}

\item{gene_name}{a selected gene for the calculation of Local Getis and Ord's Gi}

\item{k}{the number of nearest neighbours to be returned in the KNN}
}
\value{
a vector represents the Local Getis and Ord's Gi value for a selected gene across samples
}
\description{
Local Getis and Ord's Gi for measure the spatial autocorrelation of selected gene
}
