% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Louvain_clustering.R
\name{Louvain_clustering}
\alias{Louvain_clustering}
\title{This function is designed for the clustering the spots based on the spots-spots similarity}
\usage{
Louvain_clustering(
  Seurat.data,
  similarity_mat,
  setting_k = 7,
  resolution_louvain = NULL,
  max.res = 5,
  increase = 0.05
)
}
\arguments{
\item{Seurat.data}{the Seurat object generated by data_process.}

\item{similarity_mat}{a spots similarity matrix.}

\item{setting_k}{the number of clustering setting by users.}

\item{resolution_louvain}{setting the resolution for the algorithm. Provided by the user. Default setting is NULL.}

\item{max.res}{maximal resolution using for the selection of optimal resolution. Default setting is 5.}

\item{increase}{the interval between a series candidate resolution. Default setting is 0.05.}
}
\description{
The function is based on the non-spatial clustering method, Louvain's algorithm but with adaptive resolution setting
}
