% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Louvain_clustering.R
\name{Louvain_clustering_binary}
\alias{Louvain_clustering_binary}
\title{This function is designed for the clustering the spots based on the spots-spots similarity}
\usage{
Louvain_clustering_binary(
  Seurat.data,
  similarity_mat,
  setting_k = 7,
  resolution_louvain = NULL,
  min.res = 0.05,
  max.res = 5,
  max_step = 50,
  tolerance = 0
)
}
\arguments{
\item{Seurat.data}{the Seurat object generated by data_process.}

\item{similarity_mat}{a spots similarity matrix.}

\item{setting_k}{the number of clustering setting by users.}

\item{resolution_louvain}{setting the resolution for the algorithm. Provided by the user. Default setting is NULL.}

\item{min.res}{minimal resolution using for the selection of optimal resolution in binary search process. Default setting is 0.05.}

\item{max.res}{maximal resolution using for the selection of optimal resolution in binary search process. Default setting is 5.}

\item{max_step}{the maximal step for the search of resolution in Louvain algorithm.}

\item{tolerance}{the difference between ideal number of clusters and identified number of clusters. Default setting is 0.}
}
\description{
The function is based on the non-spatial clustering method, Louvain's algorithm but with adaptive resolution setting
}
