% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{patch_deal}
\alias{patch_deal}
\title{This function aims to integrate the results obtained by user own running
spatial domain methods.}
\usage{
patch_deal(res, clustering_df)
}
\arguments{
\item{res}{a list contains all the results of base SDD methods contain in EnSDD methods.}

\item{clustering_df}{a data frame contains the clustering label generated by SDD methods generated by users. The row represents the spot and the column represents the methods.}
}
\value{
a list contains clustering labels, spot-spot similarity matrix and running time.
}
\description{
This function aims to integrate the results obtained by user own running
spatial domain methods.
}
