% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_Sankey.R
\name{plot_sankey}
\alias{plot_sankey}
\title{This function is designed for the visualization of Sankey diagram.}
\usage{
plot_sankey(df, filter = 5)
}
\arguments{
\item{df}{the data frame, where row represents the sample, column represents the multiple clusters}

\item{filter}{filter the number of interaction in Sankey diagram. Default setting is 5.}
}
\description{
This function is designed for the visualization of Sankey diagram.
}
