% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_weight.R
\name{plot_weight}
\alias{plot_weight}
\title{plot weight and ARI dotplot}
\usage{
plot_weight(
  data_bar,
  sample_id,
  file_name,
  save_file = FALSE,
  label.x = 0.12,
  label.y.inc = 0.03
)
}
\arguments{
\item{data_bar}{the dataframe for plotting of dotplot.}

\item{sample_id}{the title for the dotplot.}

\item{file_name}{the file name of the saving plot.}

\item{save_file}{whether save the dotplot.}

\item{label.x}{the x-axis location of the legend.}

\item{label.y.inc}{the y-axis location of the legend.}
}
\description{
plot weight and ARI dotplot
}
