% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ct_inference.R
\name{RCTD_run}
\alias{RCTD_run}
\title{Inferring the cell type abundance in spots for SRT data}
\usage{
RCTD_run(database, CELL_MIN_INSTANCE = 20)
}
\arguments{
\item{CELL_MIN_INSTANCE}{minimum number of cells required per cell type. Default 20, can be lowered if desired.}

\item{Seurat.data}{a Seurat object contains spatial gene expression and spatial location}

\item{sc_ref}{a gene*cell scRNA-seq matrix}

\item{sc_label}{a vector represents the cell lable of cells in scRNA-seq data}
}
\value{
a matrix where row represents the spots and the column represents the cell type.
}
\description{
This function is implemented to perform deconvolution for SRT data
}
