% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_individual_clustering.R
\name{run_individual_cluster}
\alias{run_individual_cluster}
\title{Running each base clustering method individually to obtain the  base clustering results on spatially resolved transcriptomics data.}
\usage{
run_individual_cluster(
  counts_path,
  loc_path,
  img_path,
  python_env,
  n_HVG = 2000,
  n_PCA = 20,
  number.setting.by.user = 7,
  saving_results = FALSE,
  BayesSpace = TRUE,
  DR.SC = TRUE,
  SpaGCN = TRUE,
  stLearn = TRUE,
  GraphST = TRUE,
  STAGATE = TRUE,
  BayesSpace_nrep = 50000,
  DR.SC_n_SVGs = 600,
  DR.SC_latent_q = 15,
  SpaGCN_beta = 55,
  SpaGCN_alpha = 1,
  SpaGCN_p = 0.5,
  SpaGCN_l = 0.5,
  SpaGCN_res.setting = NULL,
  SpaGCN_lr = 0.05,
  SpaGCN_epoches = 200,
  stLearn_n.PCs = 30,
  stLearn_normalize_type = "physical_distance",
  GraphST_lambda1 = 10,
  GraphST_lambda2 = 1,
  GraphST_tool = "mclust",
  GraphST_radius = 50,
  GraphST_refinement = TRUE,
  GraphST_n_PCs = 20,
  STAGATE_alpha = 0
)
}
\arguments{
\item{counts_path}{the .txt path for SRT gene expression data. The file need include gene names and spots names.}

\item{loc_path}{the .txt path for meta data of the spots.}

\item{img_path}{the .tif path of H\& image of the corresponding sequencing tissue.}

\item{python_env}{the path of python environment. We recommend user construct python environment by the .yml provided by ours.}

\item{n_HVG}{the number of Highly Variation Genes (HVGs) selected by Seurat. Default setting is 2000.}

\item{n_PCA}{the number of low-dimensional of SRT gene expression. Default setting is 20.}

\item{number.setting.by.user}{the number of clustering setting by users. Default setting is 7.}

\item{saving_results}{whether to save the results of individual clustering results.}

\item{BayesSpace}{a logical variable whether to apply BayesSpace.}

\item{DR.SC}{a logical variable whether to apply DR.SC.}

\item{SpaGCN}{a logical variable whether to apply SpaGCN.}

\item{stLearn}{a logical variable whether to apply stLearn.}

\item{GraphST}{a logical variable whether to apply GraphST.}

\item{STAGATE}{a logical variable whether to apply STAGATE.}

\item{BayesSpace_nrep}{integer variable indicates the number of MCMC iteration.}

\item{DR.SC_n_SVGs}{integer variable indicates the number of spatially variable genes to be chosen. Default setting is 600.}

\item{DR.SC_latent_q}{integer variable specifies the number of latent features to be extracted. Default setting is 15.}

\item{SpaGCN_beta}{integer variable determines the area of each spot when extracting color intensity. Default value is 1.}

\item{SpaGCN_alpha}{floor variable determines the weight given to histology when calculating Euclidean distance between every two spots. Default setting is 1.}

\item{SpaGCN_p}{floor variable determines percentage of total expression contributed by neighborhoods.}

\item{SpaGCN_l}{floor variable control parameter SpaGCN_p.}

\item{SpaGCN_res.setting}{floor variable determines the resolution for Louvain's algorithm in SpaGCN.}

\item{SpaGCN_lr}{floor variable determines the learning rate in SpaGCN.}

\item{SpaGCN_epoches}{integer variable determines the maximum epoches for SpaGCN}

\item{stLearn_n.PCs}{integer variable determines the number of principal components to compute.}

\item{stLearn_normalize_type}{item determines the using spatial location (S), tissue morphological feature (M) and gene expression (E) information to normalize data in stLearn. Avariable iterms contains "weights_matrix_all","weights_matrix_pd_gd","weights_matrix_pd_md","weights_matrix_gd_md","gene_expression_correlation","physical_distance","morphological_distance". Default setting is "physical_distance".}

\item{GraphST_lambda1}{float variable determines optional weight factor to control the influence of reconstruction loss in mapping matrix learning. Default setting is 10.}

\item{GraphST_lambda2}{float variable determines optional weight factor to control the influence of contrastive loss in mapping matrix learning. Default setting is 1.}

\item{GraphST_tool}{item determines the clustering methods introduced for clustering spots. Avariable tools contain 'leiden', 'louvain', 'mclust'. Default setting is 'mclust'.}

\item{GraphST_radius}{integer variable determines the number of neighbors considered during refinement. The default is 50.}

\item{GraphST_refinement}{logical variable refines the predicted labels or not. The default is False.}

\item{GraphST_n_PCs}{integer variables determines the number PCs in the PCA. Default setting is 20.}

\item{STAGATE_alpha}{floor variable determines the weight of cell type-aware spatial neighbor network. Default setting is 0.}
}
\value{
a list contains all the results, labels and binary spot-spot similarity matrix, inferred by individual clustering methods and the times of running individual methods. The elements of list is a spots labels vactor, a matrix,  spots * spots and a time vector.
}
\description{
This function is implemented to perform individual clustering methods. The current implementation of
EnSDD integrates six state-of-the-art methods:  BayesSpace, DR-SC, GraphST, SpaGCN, stLearn and STAGATE.
These packages will be automatically installed along with EnSDD. \cr
}
